+++ Pandora/Paradigm/Primary/Functor/Function.hs
@@ -2,7 +2,7 @@
 
 module Pandora.Paradigm.Primary.Functor.Function where
 
-import Pandora.Core.Morphism ((%), (!))
+import Pandora.Core.Morphism ((%))
 import Pandora.Pattern.Category ((.), identity)
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
@@ -10,6 +10,8 @@ import Pandora.Pattern.Functor.Distributive (Distributive ((>>-)))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)))
 import Pandora.Pattern.Functor.Representable (Representable (Representation, (<#>), tabulate))
+
+infixr 2 !
 
 instance Covariant ((->) a) where
 	(<$>) = (.)
@@ -30,3 +32,7 @@ instance Representable ((->) e) where
 	type Representation ((->) e) = e
 	(<#>) = (identity %)
 	tabulate = identity
+
+{-# INLINE (!) #-}
+(!) :: a -> b -> a
+x ! _ = x