+++ Pandora/Pattern/Functor/Covariant.hs
@@ -1,8 +1,9 @@
 module Pandora.Pattern.Functor.Covariant where
 
 import Pandora.Core.Functor (type (:.), type (:=), type (<-|))
-import Pandora.Core.Morphism (fix, (!), (%))
+import Pandora.Core.Morphism (fix, (%))
 import Pandora.Pattern.Category (Category ((.), ($)))
+-- import Pandora.Paradigm.Primary.Functor.Function ((!))
 
 infixl 4 <$>, <$, $>
 infixl 3 <$$>
@@ -30,7 +31,7 @@ class Covariant (t :: * -> *) where
 	comap f x = f <$> x
 	-- | Replace all locations in the input with the same value
 	(<$) :: a -> t b -> t a
-	(<$) = comap . (!)
+	(<$) = comap . (\x _ -> x)
 	-- | Flipped version of '<$'
 	($>) :: t a -> b -> t b
 	($>) = (%) (<$)